+++ django_datalog/__init__.py
@@ -1,6 +1,6 @@
 # Package for Datalog-like fact & inference engine
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 
 default_app_config = "django_datalog.apps.DjangoDatalogConfig"